Reorganize project, add search functionality, and rename QuickBot to Pristar4.ChessEngine
This major commit involves various changes including a project reorganization that renames QuickBot to Pristar4.ChessEngine, moves several files to respective folders, and adds important functionality to the Search code segment. Changes in the project hierarchy, .csproj, and .sln files were performed to reflect the renamed project and improved structure. This commit also focused on implementing a search functionality to enhance the performance of the chess engine, right now the moves are random again but with proper shuffle method from dot net 8